// test/types.ts
// The type of ./queries.json
export type QueriessJsonEntry = {
  privateKey: string
  sendersAuthID: string
  type: string
  raw: string[]
}

// The type of ./requets.json
export type RequestsJsonEntry = {
  privateKey: string
  sendersAuthID: string
  type: string
  raw: string[]
}

// The type of ./txs.json
export type TxsJsonEntry = {
  privateKey: string
  sendersAuthID: string
  type: string
  raw: string[]
}